requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-jose==3.3.0
python-multipart==0.0.5
boto3==1.18.0
pytest==6.2.5
pytest-cov==2.12.1
python-dotenv==0.19.0
websockets==10.0

